% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bplasso.R
\name{model}
\alias{model}
\title{Compute the entire regression response model.}
\usage{
model(beta0, theta0, beta, theta, X, Z)
}
\arguments{
\item{beta0}{beta_0 coefficients}

\item{theta0}{(main effect) coefficients for Z ncol(Z)}

\item{beta}{a vectore of estimated beta coefficients each having a length   ncol(X)}

\item{theta}{theta coefficients for a single response ncol(X) by ncol(Z)}

\item{X}{N by p matrix of predictors}

\item{Z}{N by K matrix of modifying variables. The elements of Z  may represent quantitative or categorical variables, or a mixture of the two.
Categorical variables should be coded by 0-1 dummy variables: for a k-level variable, one can use either k or k-1  dummy variables.}
}
\value{
a vector of length N of the  response
}
\description{
Compute the entire regression response model.
}
