% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bplasso.R, R/myPackage-package.R
\name{PliableBVS}
\alias{PliableBVS}
\title{Fit a a Bayesian pliable lasso model over some MCMC iterations}
\usage{
PliableBVS(
  Y,
  X,
  Z,
  alpha = 0.5,
  family = c("gaussian", "binomial"),
  niter = 10000,
  burnin = 5000,
  a_rho = 1,
  b_rho = 1,
  a_zeta = 1,
  b_zeta = 1,
  num_update = 100,
  niter.update = 100,
  burnin.update = 50,
  verbose1 = FALSE,
  verbose2 = FALSE,
  lam1 = 0.1,
  lam2 = 0.1,
  rho_prior = TRUE,
  rho = 0.5,
  zeta = 0.5,
  c2 = 10^2,
  v2 = 0.1,
  update_tau = TRUE,
  option.weight.group = FALSE,
  option.update = "global",
  lambda2_update = NULL,
  method = NULL
)
}
\arguments{
\item{X}{N by p matrix of predictors}

\item{Z}{N by K matrix of modifying variables. The elements of Z  may represent quantitative or categorical variables, or a mixture of the two.
Categorical variables should be coded by 0-1 dummy variables: for a k-level variable, one can use either k or k-1  dummy variables.}

\item{family}{response type- either "gaussian", "binomial". In the binomial case, y should be 0s and 1s.}

\item{y}{N by D matrix  of responses. The X and Z variables are centered in the function. We recommend that X and Z also be standardized before the call}
}
\value{
a list of all the posterior objects
}
\description{
Fit a a Bayesian pliable lasso model over some MCMC iterations
}
