% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bplasso.R
\name{predict.Bpliable}
\alias{predict.Bpliable}
\title{Compute predicted values from a fitted Bpliable  object.
 Make predictions from a fitted Bpliable model}
\usage{
\method{predict}{Bpliable}(object, X, Z, prob = 0.5)
}
\arguments{
\item{object}{object returned from a call to Bpliable}

\item{X}{N by p matrix of predictors}

\item{Z}{N by nz matrix of modifying variables. These may be observed or
the predictions from a supervised learning algorithm that predicts z from
test features x  and possibly other features.}

\item{prob}{threshold for binomial prediction}
}
\value{
predicted response or class
}
\description{
Compute predicted values from a Bpliable object.
 Make predictions from a fitted Bpliable model
}
